Move "pkg/binauthz" package into "pkg/kritis"
